requirements.txt
django==1.10
django-rest-framework
django-filter==1.1.0